movie_manager/etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["horw horw <horw@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
